+++ components/Navigation/Settings.tsx
@@ -1,11 +1,15 @@
 import { TouchableOpacity, Text, View } from "react-native";
-import React from "react";
-import { SettingsPropsType } from "../Types/typeconfig";
-import styles from "../Styles/styles";
+import React, { useContext } from "react";
+import { SettingsPropsType } from "../../Types/typeconfig";
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import ThemeContext from '../../theme/ThemeContext';
+import LightStyles from '../Styles/LightStyles';
+import DarkStyles from '../Styles/DarkStyles';
 
 const Settings = (props: SettingsPropsType) => {
   const { navigation } = props;
+  const { theme } = useContext(ThemeContext);
+  const styles = theme === 'light' ? LightStyles : DarkStyles;
 
   const handleChangeEmail = () => {
     // write the logic to change email
