+++ components/inputs/index.tsx
@@ -1,6 +1,5 @@
 import SalaryForm from "./SalaryForm";
 import TransactionView from "./TransactionView";
-import NewTransaction from "./NewTransaction";
 import AccountSettings from "./AccountSettings";
 
-export { SalaryForm, TransactionView, NewTransaction, AccountSettings };
+export { SalaryForm, TransactionView, AccountSettings };
