--- inputs/TransactionForm.tsx
@@ -1,72 +1,0 @@
-import React, { useState } from 'react';
-import { Text, View, TextInput, TouchableOpacity, Platform } from 'react-native';
-import DateTimePicker from '@react-native-community/datetimepicker';
-import styles from '../styles';
-
-export default function TransactionForm() {
-    const [date, setDate] = useState(new Date());
-    const [showDatePicker, setShowDatePicker] = useState(false);
-
-    const [form, setForm] = useState({
-        description: '',
-        amount: '',
-        date: '',
-    });
-
-    const handleDateChange = (event: any, selectedDate?: Date) => {
-        const currentDate = selectedDate || date;
-        //android = show date picker - due to having set the date if I was going to select the date again the date picker would not show
-        //ios doesnt need such logic
-        if (Platform.OS === 'android') {
-            setShowDatePicker(false);
-        }
-        setDate(currentDate);
-    };
-
-    const handleTransaction = () => {
-        // Implement your transaction logic here
-    };
-
-    return (
-        <View style={styles.container}>
-            <View style={styles.inputContainer}>
-                <TextInput
-                    style={styles.input}
-                    onChangeText={(description) => setForm((prevState) => ({ ...prevState, description }))}
-                    placeholder="Description"
-                />
-            </View>
-            <View style={styles.inputContainer}>
-                <TextInput
-                    style={styles.input}
-                    onChangeText={(amount) => setForm((prevState) => ({ ...prevState, amount }))}
-                    placeholder="Amount"
-                />
-            </View>
-                <TouchableOpacity
-    style={styles.inputContainer}
-    onPress={() => setShowDatePicker(true)}
->
-    <Text style={styles.input}>
-        {date.toDateString()}
-    </Text>
-</TouchableOpacity>
-            {showDatePicker && (
-                <DateTimePicker
-                    testID="dateTimePicker"
-                    value={date}
-                    mode="date"
-                    is24Hour={true}
-                    display="default"
-                    onChange={handleDateChange}
-                />
-            )}
-            <TouchableOpacity
-                style={styles.button}
-                onPress={handleTransaction}
-            >
-                <Text style={styles.buttonText}>Submit</Text>
-            </TouchableOpacity>
-        </View>
-    );
-}
