--- inputs/index.tsx
@@ -1,8 +1,0 @@
-import RegistrationForm from "./RegistrationForm";
-import LoginForm from "./LoginForm";
-import SalaryForm from "./SalaryForm";
-import TransactionView from "./TransactionView";
-import TransactionForm from "./TransactionForm";
-import AccountSettings from "./AccountSettings";
-
-export { RegistrationForm, LoginForm, SalaryForm, TransactionView, TransactionForm, AccountSettings };
